Fix CIDR to World Entity Conversion Bug
Previously if either the "0.0.0.0/0" CIDR (hereafter "world-ipv4")
or a world-ipv6 CIDR was present the selector creation logic
for CIDRs would also add the "reserved:world" entity to the
selector labels generated by the CIDR.

This does not work when Cilium is dual-stack mode as "reserved:world"
covers both IPv4 and IPv6, so both CIDRs should be present before such
an addition is made. Moreover, dual-stack mode lacked any mechanism to
distinguish world-ipv4 and world-ipv6 from another. This bug has prevented
the possibility of managing world-ipv4 and world-ipv6 separately from one
another in the policy engine.

Separate identities (and labels, etc) for world-ipv4 and world-ipv6 now exist
so they can be distinguished in userspace and in the datapath.

Signed-off-by: Nate Sweet <nathanjsweet@pm.me>